keep benchmark functions in seperate modules
